// src/app/models/documentosfree/PostDocumentosFreeModel.ts
export class PostDocumentosFreeModel{
    razaosocial: string = '';
    cnpj: string = '';
    setor: string = '';
    cenario: string = '';
    datainicio: string = '';
    vigencia: string = '';
    endereco: string = '';
    servico: string = '';
    funcoesenvolvidas: string = '';
    equipamentos: string = '';
    responsavel: string = '';
    emitentes: string = '';
    cipa: string = '';
    brigada: string = '';
    resgate: string = '';
    ambulancia: string = '';
    seguranca: string = '';
    local_espaco_confinado: string = '';
    autorizados_vigia: string = '';
    autorizados_supervisor: string = '';
    autorizados_resgate: string = '';
    idEmpresa: string = '';
	
}